Jjp 135/fix duplicate packages after merge (#28)
* JJP-135: Fixed the problem with duplicated packages in main

* JJP-135: Remove unnecessary log files
