--- auth_app/src/main/java/com.blankfactor.auth/config/SecurityConfig.java
@@ -1,34 +1,0 @@
-package com.blankfactor.auth.config;
-
-import com.blankfactor.auth.config.JwtAuthenticationFilter;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.config.http.SessionCreationPolicy;
-
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Slf4j
-public class SecurityConfig {
-
-    private final AuthenticationProvider authenticationProvider;
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        log.info("Configuring SecurityFilterChain.");
-        http
-                .csrf(csrf -> csrf.disable())
-                .authorizeHttpRequests(authorize -> authorize
-                        .requestMatchers("/api/v1/auth/**").permitAll()
-                        .anyRequest().authenticated()
-                )
-                .sessionManagement(session -> session
-                        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                )
-                .authenticationProvider(authenticationProvider)
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
-        log.debug("SecurityFilterChain configured.");
-        return http.build();
-    }
-
-}
