--- auth_app/src/main/java/com.blankfactor.auth/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.blankfactor.auth.repository;
-
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    @Query("SELECT u FROM User u WHERE u.email = :identifier OR u.username = :identifier")
-    Optional<User> findByEmailOrUsername(@Param("identifier") String identifier);
-}
