+++ auth_app/src/main/java/com/blankfactor/auth/config/AppConfig.java
@@ -1,13 +1,28 @@
 package com.blankfactor.auth.config;
 
+import com.blankfactor.auth.repository.UserRepository;
+import lombok.extern.log4j.Log4j2;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Configuration
+@Log4j2
 public class AppConfig {
+
+    private final UserRepository userRepository;
+
+    public AppConfig(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
     @Bean
     public PasswordEncoder passwordEncoder() {
@@ -17,6 +32,36 @@ public class AppConfig {
     @Bean
     public ModelMapper modelMapper() {
         return new ModelMapper();
+    }
+
+    @Bean
+    UserDetailsService userDetailsService() {
+        log.info("Initializing UserDetailsService bean.");
+        return identifier -> {
+            log.debug("Loading user by email or username: {}", identifier);
+            return userRepository.findByEmailOrUsername(identifier)
+                    .orElseThrow(() -> {
+                        log.warn("User not found with email: {}", identifier);
+                        return new UsernameNotFoundException("User not found");
+                    });
+        };
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
+        log.info("Initializing AuthenticationManager bean.");
+        return config.getAuthenticationManager();
+    }
+
+    @Bean
+    AuthenticationProvider authenticationProvider() {
+        log.info("Initializing AuthenticationProvider bean.");
+        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
+
+        authProvider.setUserDetailsService(userDetailsService());
+        authProvider.setPasswordEncoder(passwordEncoder());
+
+        return authProvider;
     }
 
 }
