+++ auth_app/src/main/java/com/blankfactor/auth/controller/UserController.java
@@ -4,7 +4,7 @@ import com.blankfactor.auth.entity.User;
 import com.blankfactor.auth.entity.dto.exp.UserResponse;
 import com.blankfactor.auth.service.UserService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/users")
 @RestController
 @RequiredArgsConstructor
-@Slf4j
+@Log4j2
 public class UserController {
 
     private final UserService userService;
