+++ auth_app/src/main/java/com/blankfactor/auth/exception/custom/InvalidCredentialsException.java
@@ -1,7 +1,7 @@
 package com.blankfactor.auth.exception.custom;
 
 public class InvalidCredentialsException extends RuntimeException {
-    public InvalidCredentialsException(String message) {
-        super(message);
-    }
+  public InvalidCredentialsException(String message) {
+    super(message);
+  }
 }
