+++ auth_app/src/main/java/com/blankfactor/auth/exception/custom/user/UserNotVerifiedException.java
